backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
PyYAML
gitpython
requests
networkx
pymilvus
ollama
tree-sitter
tree-sitter-python
